backend/requirements.txt
django 
pyodbc
django-mssql-backend 
djangorestframework
dotenv
mssql-django
django-cors-headers
gunicorn
django-cors